Refactor Transcription Module Names
