+++ src/openbench/pipeline/transcription/__init__.py
@@ -6,9 +6,9 @@ from .common import TranscriptionOutput
 from .transcription_groq import GroqTranscriptionConfig, GroqTranscriptionPipeline
 from .transcription_whisperkitpro import WhisperKitProTranscriptionConfig, WhisperKitProTranscriptionPipeline
 from .whisperkit import WhisperKitTranscriptionConfig, WhisperKitTranscriptionPipeline
-from .transcription_openai_boosting import OpenAITranscriptionPipelineConfig, OpenAITranscriptionPipeline
-from .transcription_deepgram_boosting import DeepgramTranscriptionPipelineConfig, DeepgramTranscriptionPipeline
-from .transcription_nemo_boosting import NeMoTranscriptionPipelineConfig, NeMoTranscriptionPipeline
+from .transcription_opena import OpenAITranscriptionPipelineConfig, OpenAITranscriptionPipeline
+from .transcription_deepgram import DeepgramTranscriptionPipelineConfig, DeepgramTranscriptionPipeline
+from .transcription_nemo import NeMoTranscriptionPipelineConfig, NeMoTranscriptionPipeline
 
 
 __all__ = [
