+++ src/openbench/pipeline/transcription/transcription_openai.py
